+++ src/jvmMain/kotlin/net/dinkla/raytracer/objects/acceleration/kdtree/builder/TestBuilder.kt
@@ -86,7 +86,7 @@ class TestBuilder : TreeBuilder {
 
                 fun max(): Split {
                     val s = Split(null)
-                    s.sah = java.lang.Float.POSITIVE_INFINITY.toDouble()
+                    s.sah = Double.POSITIVE_INFINITY
                     return s
                 }
             }
